+++ config/achievements.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'models' => [
+        'achievement' => \Laravel\Achievements\AchievementModel::class,
+        'criteria' => \Laravel\Achievements\AchievementCriteriaModel::class,
+    ],
+];