+++ tests/AchievementsManagerTest.php
@@ -3,10 +3,10 @@
 namespace Laravel\Achievements\Tests;
 
 use Illuminate\Support\Facades\Event;
-use Laravel\Achievements\AchievementsStorage;
 use Laravel\Achievements\Events\AchievementsCompleted;
+use Laravel\Achievements\Tests\Stubs\OverridenAchievement;
+use Laravel\Achievements\Tests\Stubs\OverridenAchievementCriteria;
 use Laravel\Achievements\Tests\Stubs\User;
-use Zurbaev\Achievements\AchievementCriteria;
 use Zurbaev\Achievements\AchievementCriteriaChange;
 use Zurbaev\Achievements\AchievementsManager;
 
@@ -28,7 +28,7 @@ class AchievementsManagerTest extends TestCase
 
         $this->seedAchievements();
         $this->user = User::create(['name' => 'John Doe']);
-        $this->manager = new AchievementsManager(new AchievementsStorage());
+        $this->manager = app(AchievementsManager::class);
     }
 
     public function testCriteriaUpdated()
@@ -73,6 +73,40 @@ class AchievementsManagerTest extends TestCase
         $this->user = $this->user->fresh();
         $this->assertSame(1, count($this->user->achievements));
         $this->assertSame(10, $this->user->achievementPoints());
+    }
+
+    public function testItWorksWithReplacedModels()
+    {
+        $this->app['config']->set('achievements.models.achievement', OverridenAchievement::class);
+        $this->app['config']->set('achievements.models.criteria', OverridenAchievementCriteria::class);
+
+        Event::fake();
+
+        AchievementsManager::registerHandler('reach_level', function ($owner, $criteria, $achievement, $data) {
+            return new AchievementCriteriaChange(intval($data['value'] ?? 1), AchievementCriteriaChange::PROGRESS_ACCUMULATE);
+        });
+
+        $this->assertSame(0, count($this->user->achievements));
+        $this->assertSame(0, $this->user->achievementPoints());
+
+        $count = $this->manager->updateAchievementCriterias(
+            $this->user, 'reach_level', [
+                'value' => 10,
+            ]
+        );
+
+        Event::assertDispatched(AchievementsCompleted::class);
+
+        $this->assertSame(3, $count);
+
+        $this->user = $this->user->fresh();
+        $this->assertSame(1, count($this->user->achievements));
+        $this->assertSame(10, $this->user->achievementPoints());
+
+        $achievement = $this->user->achievements->first();
+        $this->assertInstanceOf(OverridenAchievement::class, $achievement);
+        $criteria = $this->user->achievementCriterias->first();
+        $this->assertInstanceOf(OverridenAchievementCriteria::class, $criteria);
     }
 
     public function testProgressDataIsSavedToDb()