+++ tests/Stubs/OverridenAchievement.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Laravel\Achievements\Tests\Stubs;
+
+use Laravel\Achievements\AchievementModel;
+
+class OverridenAchievement extends AchievementModel
+{
+}