+++ tests/Stubs/OverridenAchievementCriteria.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Laravel\Achievements\Tests\Stubs;
+
+use Laravel\Achievements\AchievementCriteriaModel;
+
+class OverridenAchievementCriteria extends AchievementCriteriaModel
+{
+}